First working version of the script that will generate time code for all clips in the active timeline.
It will create a new folder called "Adjusted TC for ______" where _____ is the name of the sequence. Sub folders will be created for each track with copies of the clips placed in each folder. The clips timecode will be changed to reflect the location of the clips in the timeline.  The script outputs various information to the console including what tracks it found, the locations of the clips.  If a clip is placed multiple times on the timeline it will be skipped after the first occurrence is discovered. A warning will be printed to the console.  Both the "Angle" and "Camera #" are set to the track name for the clip.

Limitations
- Clips are assumed to be uncut on the timeline. Currently doesn't warn if this is the case.
- Folders will be created for the audio tracks from video files. Audio tracks from video files are skipped as the timecode is already set for the clip.
- Timecode for clips will be changed regardless of what the timecode is set.
- Clips will refuse to sync if their file creation/modification date is different from each other regardless of the timecode set.
- Audio clips are duplicated, but new media pool items are created for video clips.  If a video clips's timecode is modified, it will cause the clip to go offline. This doesn't happen for a audio clip.